refactor: Filter 컴포넌트 구조 개선
- Filter 컴포넌트를 독립적인 디렉토리로 이동
- Filter 컴포넌트 경로 변경에 따른 import 경로 수정
- mockProducts를 props로 전달하도록 변경
