+++ src/app/products/_components/filter/constants.ts
@@ -1,0 +1,10 @@
+export const ALLOWED_FILTERS = ['Color'];
+
+export const COLOR_MAPPING: Record<string, string> = {
+  'Natural Titanium': '#E3D5C9',
+  'Blue Titanium': '#4B688A',
+  'Phantom Black': '#000000',
+  Cream: '#F5E6D3',
+  'Space Gray': '#4A4A4A',
+  Silver: '#E3E4E5',
+};
