+++ src/app/products/_components/index.ts
@@ -1,2 +1,2 @@
 export { default as CategoryList } from './CategoryList';
-export { default as Filter } from './Filter';
+export { default as Filter } from './filter/Filter';
